OT120-34 creo el metodo para actualizar una categoria por id, y su endpoint.
