+++ src/main/java/com/alkemy/ong/OngApplication.java
@@ -1,6 +1,5 @@
 package com.alkemy.ong;
 
-import com.alkemy.ong.model.response.CategoryDtoMapping;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
