+++ src/main/java/com/alkemy/ong/controller/CategoryController.java
@@ -1,7 +1,9 @@
 package com.alkemy.ong.controller;
 
 import com.alkemy.ong.model.request.CategoryDto;
-import com.alkemy.ong.model.entity.Category;
+import com.alkemy.ong.entity.Category;
+import com.alkemy.ong.dto.CategoryRequest;
+import com.alkemy.ong.dto.CategoryResponse;
 import com.alkemy.ong.service.CategoryServiceImpl;
 import com.alkemy.ong.service.abstraction.IDeleteCategoryService;
 import com.fasterxml.jackson.databind.introspect.TypeResolutionContext.Empty;
@@ -9,11 +11,7 @@ import javax.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RestController;
-
-import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
@@ -40,6 +38,13 @@ public class CategoryController {
     @GetMapping("/categories/{id}")
     public ResponseEntity<Category> getOne(@PathVariable long id) throws EntityNotFoundException {
         return ResponseEntity.status(HttpStatus.OK).body(categoryService.getCategory(id));
+    }
+
+    @PutMapping("/categories/{id}")
+    public ResponseEntity<CategoryResponse> update(@PathVariable long id, @RequestBody CategoryRequest categoryDto)
+            throws EntityNotFoundException {
+        CategoryResponse updatedCategory = categoryService.update(id, categoryDto);
+        return ResponseEntity.ok().body(updatedCategory);
     }
 
 }
