+++ src/main/java/com/alkemy/ong/dto/CategoryRequest.java
@@ -1,0 +1,16 @@
+package com.alkemy.ong.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.springframework.stereotype.Component;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class CategoryRequest {
+
+    public String name;
+    public String description;
+    public String image;
+}
