+++ src/main/java/com/alkemy/ong/entity/Activity.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.model.entity;
+package com.alkemy.ong.entity;
 
 import java.sql.Timestamp;
 import javax.persistence.Column;
@@ -14,28 +14,28 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 
-@Entity
-@Table(name = "MEMBERS")
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class Member {
+@Entity
+@Table(name = "ACTIVITIES")
+public class Activity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Setter(AccessLevel.NONE)
-  @Column(name = "MEMBERS_ID")
+  @Column(name = "ACTIVITIES_ID")
   private long id;
 
   @Column(name = "NAME", nullable = false)
   private String name;
 
+  @Column(name = "CONTENT", nullable = false)
+  private String content;
+
   @Column(name = "IMAGE", nullable = false)
   private String image;
-
-  @Column(name = "DESCRIPTION")
-  private String description;
 
   @Column(name = "TIMESTAMPS")
   @CreationTimestamp
