+++ src/main/java/com/alkemy/ong/entity/Contact.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.model.entity;
+package com.alkemy.ong.entity;
 
 import lombok.Data;
 import lombok.Getter;
