+++ src/main/java/com/alkemy/ong/entity/Member.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.model.entity;
+package com.alkemy.ong.entity;
 
 import java.sql.Timestamp;
 import javax.persistence.Column;
@@ -14,32 +14,32 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 
+@Entity
+@Table(name = "MEMBERS")
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-@Entity
-@Table(name = "CATEGORIES")
-public class Category {
+public class Member {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "CATEGORIES_ID")
   @Setter(AccessLevel.NONE)
+  @Column(name = "MEMBERS_ID")
   private long id;
 
   @Column(name = "NAME", nullable = false)
   private String name;
 
+  @Column(name = "IMAGE", nullable = false)
+  private String image;
+
   @Column(name = "DESCRIPTION")
   private String description;
 
-  @Column(name = "IMAGE")
-  private String image;
-
+  @Column(name = "TIMESTAMPS")
   @CreationTimestamp
-  @Column(name = "TIMESTAMP")
-  private Timestamp timestamp;
+  private Timestamp timestamps;
 
   @Column(name = "SOFT_DELETE")
   private boolean softDelete;
