+++ src/main/java/com/alkemy/ong/mapper/CategoryMapper.java
@@ -1,7 +1,9 @@
 package com.alkemy.ong.mapper;
 
-import com.alkemy.ong.model.entity.Category;
+import com.alkemy.ong.entity.Category;
 import com.alkemy.ong.model.request.CategoryDto;
+import com.alkemy.ong.dto.CategoryRequest;
+import com.alkemy.ong.dto.CategoryResponse;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -14,5 +16,13 @@ public class CategoryMapper {
     
     public CategoryDto categoryToCategoryDto(Category entity){
         return modelMapper.map(entity, CategoryDto.class);
-    }   
+    }
+
+    public Category categoryRequest2Entity(CategoryRequest categoryDto) {
+        return modelMapper.map(categoryDto, Category.class);
+    }
+
+    public CategoryResponse category2Dto(Category entity) {
+        return modelMapper.map(entity, CategoryResponse.class);
+    }
 }
