+++ src/main/java/com/alkemy/ong/repository/IActivityRepository.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.repository;
 
-import com.alkemy.ong.model.entity.Activity;
+import com.alkemy.ong.entity.Activity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
