+++ src/main/java/com/alkemy/ong/repository/ICommentRepository.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.repository;
 
-import com.alkemy.ong.model.entity.Comment;
+import com.alkemy.ong.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
