+++ src/main/java/com/alkemy/ong/repository/IContactRepository.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.repository;
 
-import com.alkemy.ong.model.entity.Contact;
+import com.alkemy.ong.entity.Contact;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
