+++ src/main/java/com/alkemy/ong/repository/IMemberRepository.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.repository;
 
-import com.alkemy.ong.model.entity.Member;
+import com.alkemy.ong.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
