+++ src/main/java/com/alkemy/ong/repository/ISlideRepository.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.repository;
 
-import com.alkemy.ong.model.entity.Slide;
+import com.alkemy.ong.entity.Slide;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
