+++ src/main/java/com/alkemy/ong/repository/ITestimonialRepository.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.repository;
 
-import com.alkemy.ong.model.entity.Testimonial;
+import com.alkemy.ong.entity.Testimonial;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
