+++ src/main/java/com/alkemy/ong/service/AuthenticationService.java
@@ -2,7 +2,7 @@ package com.alkemy.ong.service;
 
 
 import com.alkemy.ong.common.JwtUtil;
-import com.alkemy.ong.model.entity.User;
+import com.alkemy.ong.entity.User;
 import com.alkemy.ong.model.request.LoginRequest;
 import com.alkemy.ong.model.response.TokenDto;
 import com.alkemy.ong.exception.InvalidCredentialsException;
