+++ src/main/java/com/alkemy/ong/service/MemberServiceImpl.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.service;
 
-import com.alkemy.ong.model.entity.Member;
+import com.alkemy.ong.entity.Member;
 import com.alkemy.ong.repository.IMemberRepository;
 import com.alkemy.ong.service.abstraction.IDeleteMembersService;
 import java.util.Optional;
