+++ src/main/java/com/alkemy/ong/service/abstraction/IRoleService.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.service.abstraction;
 
-import com.alkemy.ong.model.entity.Role;
+import com.alkemy.ong.entity.Role;
 
 public interface IRoleService {
