+++ src/test/java/com/alkemy/ong/integration/user/DeleteUserIntegrationTest.java
@@ -6,7 +6,7 @@ import static org.mockito.Mockito.when;
 
 import com.alkemy.ong.common.AbstractBaseIntegrationTest;
 import com.alkemy.ong.config.security.ApplicationRole;
-import com.alkemy.ong.model.entity.User;
+import com.alkemy.ong.entity.User;
 import com.alkemy.ong.model.response.ErrorResponse;
 import java.util.Optional;
 import org.junit.Test;
